+++ shared/protocol/build.rs
@@ -23,6 +23,7 @@ fn compile_protos() {
                 "proto/reovim/v2/module.proto",
                 "proto/reovim/v2/server.proto",
                 "proto/reovim/v2/notification.proto",
+                "proto/reovim/v2/syntax.proto",
             ],
             &["proto"],
         )